requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
astor==0.8.1
attrs==23.2.0
bandit==1.7.6
certifi==2023.11.17
charset-normalizer==3.3.2
darglint==1.8.1
distro==1.9.0
Django==5.0.1
djangorestframework==3.14.0
docutils==0.20.1
eradicate==2.3.0
exceptiongroup==1.2.0
flake8==6.1.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==23.12.2
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
pbr==6.0.0
pep8-naming==0.13.3
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.3.0
pyflakes==3.1.0
Pygments==2.17.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
restructuredtext-lint==1.4.0
rich==13.7.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sqlparse==0.4.4
stevedore==5.1.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
wemake-python-styleguide==0.18.0
